backend/requirements.txt
fastapi==0.116.1
uvicorn==0.30.1
sqlalchemy==2.0.41
passlib[bcrypt]==1.7.4
python-jose==3.5.0
pydantic[email]==2.11.7
python-dotenv==1.0.1
openai==0.28.0
httpx==0.28.1
python-multipart==0.0.20
